ENH: Add additional analyses
Add additional analyses to provide a more comprehensive picture of the
author and citation diversity, equity and inclusiveness:
- Report the number of instances a given author from the references
  citation file is listed as first/last.
- Report the number of different first/last authors.
- Report the gender distribution of an author citation file.
- Report the number of instances authors are cited in the references.
- Report the number of instances authors are cited as first/last in the
  references.
- Make the `missed` first names be a dictionary instead of a set in
  order to save the number of instances the missed first names were
  missed (i.e. how many times they are present in the relevant author
  list).
- Serialize the extracted features (gender, counts).

Refactor the script for an improved code encapsulation.

Given that the number of serialized features has increased, prefer
specifying the output folder instead of requiring input filenames as
program arguments.

Remove unused global variables.

Rename some global variables for the sake of consistency.

Sort functions in a more conventional way: make the parsing and `main`
functions come last.

Sort the imported packages.

Document the functions.

Provide some minimal description in the `README` file of the features
the tool is able to extract.